Week1_RAG_Basic/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-openai
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
pypdf
streamlit
requests
python-dotenv